requirements.txt
numpy
pandas
scikit-learn
matplotlib
tensorflow
pyyaml
tqdm
joblib
